Cleanup and better features

+++ vtils/ipc/shared_memory.py
@@ -17,26 +17,27 @@ Output: shared_memory_array
 
 
 class shared_memory_array:
-    def __init__(
-        self,
-        name: str,
-        init_array: np.array = None,  # initial array values to use
-        shape: np.shape = None,
-        dtype: np.dtype = None,
-    ):
+    def __init__(self, name: str, init_array: np.array):  # initial array values to use
         self.shm = None
         self.val = None
-        shape = init_array.shape
-        dtype = init_array.dtype
+
+        if init_array is not None:
+            shape = init_array.shape
+            dtype = init_array.dtype
 
         try:
             self.access_shared_memory(memory_name=name, shape=shape, dtype=dtype)
             print(f"Vtils:> Shared memory ({name}) found.")
-        except:
-            print(
-                f"Vtils:> Shared memory ({name}) not found. Created a new shared memory using init_array"
-            )
-            self.register_shared_memory(memory_name=name, memory_value=init_array)
+        except FileNotFoundError:
+            if init_array is None:
+                raise ValueError(
+                    f"Vtils:> Shared memory ({name}) not found and no init_array provided."
+                )
+            else:
+                print(
+                    f"Vtils:> Shared memory ({name}) not found. Created a new shared memory using init_array"
+                )
+                self.register_shared_memory(memory_name=name, memory_value=init_array)
 
     def register_shared_memory(self, memory_name, memory_value):
         shm = shared_memory.SharedMemory(
@@ -56,36 +57,47 @@ class shared_memory_array:
         self.val = val_shared
 
     def close_link(self):
-        # Close access to the shared memory from this instance.
-        self.shm.close()
+        if self.shm is not None:
+            # Close access to the shared memory from this instance.
+            self.shm.close()
+            print(f"Vtils:> Shared memory ({self.name}) closed")
 
     def delete_memory(self):
-        # Request that the underlying shared memory block be destroyed. Call only once
-        self.shm.unlink()
+        if self.shm is not None:
+            # Request that the underlying shared memory block be destroyed. Call only once
+            self.shm.unlink()
+            print(f"Vtils:> Shared memory ({self.name}) deleted")
+            self.shm = None
 
     @property
     def name(self):
         if self.shm is not None:
-
             return self.shm.name
+
+    def __del__(self):
+        self.close_link()
 
 
 if __name__ == "__main__":
     print(HELP)
     user_data = np.array([1, 1, 2, 3, 5, 8.0])
+    print(f"Original user data: {user_data}")
 
     # share data on shared memory
     shared_user_data = shared_memory_array(name="shared_mem_name", init_array=user_data)
 
     # access data on shared memory
-    access_user_data = shared_memory_array(name="shared_mem_name", init_array=user_data)
+    for i in range(3):
+        access_user_data = shared_memory_array(
+            name="shared_mem_name", init_array=user_data
+        )
 
-    # Check data
-    print(f"Original user data: {user_data}")
-    print(f"Data accessed over shared memory: {access_user_data.val}")
+        # Check accessed data
+        print(f"\t[Loop{i}]: Data accessed over shared memory: {access_user_data.val}")
 
-    # release shared memory link
-    shared_user_data.close_link()
+        # release shared memory link
+        shared_user_data.close_link()
+
     access_user_data.close_link()
 
     # Request removal of shared memory. Only one process needs to call it
